requirements.txt
pyside2
psutil
pyaml